+++ vm/src/hint_processor/builtin_hint_processor/secp/ec_utils.rs
@@ -119,12 +119,14 @@ Implements hint:
     value = slope = ec_double_slope(point=(x, y), alpha=0, p=SECP_P)
 %}
 */
+#[allow(clippy::too_many_arguments)]
 pub fn compute_doubling_slope(
     vm: &mut VirtualMachine,
     exec_scopes: &mut ExecutionScopes,
     ids_data: &HashMap<String, HintReference>,
     ap_tracking: &ApTracking,
     point_alias: &str,
+    pack_prime: &BigUint,
     secp_p: &BigInt,
     alpha: &BigInt,
 ) -> Result<(), HintError> {
@@ -132,7 +134,14 @@ pub fn compute_doubling_slope(
     //ids.point
     let point = EcPoint::from_var_name(point_alias, vm, ids_data, ap_tracking)?;
 
-    let value = ec_double_slope(&(point.x.pack86(), point.y.pack86()), alpha, secp_p)?;
+    let value = ec_double_slope(
+        &(
+            point.x.pack86_for_prime(pack_prime),
+            point.y.pack86_for_prime(pack_prime),
+        ),
+        alpha,
+        secp_p,
+    )?;
     exec_scopes.insert_value("value", value.clone());
     exec_scopes.insert_value("slope", value);
     Ok(())